Push temporary changes.

+++ lib/utils.h
@@ -11,7 +11,7 @@
 #include "linmath.h"
 #include "stb_image.h"
 
-#define M_PI 3.14159265358979323846264338327950288
+#define M_PI (float)3.14159265358979323846264338327950288
 
 char *
 read_file(const char * path_file);
